+++ help-desk-api/Models/Entities/Setup/EmailConfigurationModel.cs
@@ -3,7 +3,7 @@ using System.ComponentModel.DataAnnotations.Schema;
 namespace Models.Entities.Setup
 {
     [Table("EmailConfiguration", Schema = "setup")]
-    public class EmailConfigurationModel : BaseEntity<long>
+    public class EmailConfigurationModel : BaseEntity<int>
     {
         public string UserName { get; set; }
         public string Password { get; set; }
